chore: created headers and source for PathTrackingController class
